backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
pymongo==4.6.2
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
google-generativeai==0.4.1
cloudinary==1.40.0
